@@ -23,21 +23,20 @@
     "joi-validation-strategy": "^0.3.3",
     "lodash": "^4.9.0",
     "lodash.assign": "^4.0.7",
-    "material-ui": "^0.14.4",
+    "material-ui": "^0.15.0",
     "q": "^1.4.1",
-    "react": "^0.14.8",
-    "react-addons-css-transition-group": "^0.14.8",
-    "react-addons-transition-group": "^0.14.8",
+    "react": "^15.1.0",
+    "react-addons-css-transition-group": "^15.1.0",
+    "react-addons-transition-group": "^15.1.0",
     "react-dnd": "^2.1.4",
     "react-dnd-touch-backend": "^0.2.5",
-    "react-dom": "^0.14.8",
-    "react-flipcard": "^0.2.1",
+    "react-dom": "^15.1.0",
     "react-google-charts": "^1.0.0",
     "react-redux": "^4.4.2",
     "react-router": "^2.0.1",
     "react-router-redux": "^4.0.2",
     "react-slider": "^0.6.1",
-    "react-tap-event-plugin": "^0.2.2",
+    "react-tap-event-plugin": "^1.0.0",
     "react-validation-mixin": "^5.3.4",
     "redux": "^3.4.0",
     "redux-actions": "^0.9.1",
