@@ -7,8 +7,8 @@ import * as React from 'react'
 var { Component } = React
 var { Chart } = require('../../../forked/react-google-charts/Chart.js')
 // var { Chart } = require('react-google-charts')
-import IconButton = require('material-ui/lib/icon-button')
-import FontIcon = require('material-ui/lib/font-icon')
+import IconButton from 'material-ui/IconButton'
+import FontIcon from 'material-ui/FontIcon'
 import {
     ChartParms,
     ChartSettings,
