@@ -1,10 +1,9 @@
 'use strict';
 const React = require('react');
 var { Component } = React;
-const TextField = require('material-ui/lib/text-field');
-const CardText = require('material-ui/lib/card/card-text');
-const CardActions = require('material-ui/lib/card/card-actions');
-const RaisedButton = require('material-ui/lib/raised-button');
+const TextField_1 = require('material-ui/TextField');
+const Card_1 = require('material-ui/Card');
+const RaisedButton_1 = require('material-ui/RaisedButton');
 class BasicForm extends React.Component {
     constructor(...args) {
         super(...args);
@@ -34,15 +33,15 @@ class BasicForm extends React.Component {
             return (React.createElement("div", {className: "textfieldwrapper", style: {
                 display: display,
                 marginRight: "5px"
-            }, key: element.index}, React.createElement(TextField, React.__spread({ref: node => { basicform.textFields[element.index] = node; }}, attributes))));
+            }, key: element.index}, React.createElement(TextField_1.default, React.__spread({ref: node => { basicform.textFields[element.index] = node; }}, attributes))));
         });
-        return (React.createElement("form", {onSubmit: basicform.submit, action: "javascript:void(0)"}, React.createElement(CardText, null, basicform.props.completionMessage
+        return (React.createElement("form", {onSubmit: basicform.submit, action: "javascript:void(0)"}, React.createElement(Card_1.CardText, null, basicform.props.completionMessage
             ? React.createElement("p", {style: { color: "green" }}, basicform.props.completionMessage)
             : null, basicform.props.warningMessage
             ? React.createElement("p", {style: { color: "orange" }}, basicform.props.warningMessage)
             : null, basicform.props.errorMessage
             ? React.createElement("p", {style: { color: "red" }}, basicform.props.errorMessage)
-            : null, children), React.createElement(CardActions, null, React.createElement(RaisedButton, {type: "submit", label: basicform.props.submitButtonLabel, className: "button-submit", primary: true}))));
+            : null, children), React.createElement(Card_1.CardActions, null, React.createElement(RaisedButton_1.default, {type: "submit", label: basicform.props.submitButtonLabel, className: "button-submit", primary: true}))));
     }
 }
 exports.BasicForm = BasicForm;
