@@ -1,7 +1,7 @@
 'use strict';
 const React = require('react');
 var { Component } = React;
-const GridList = require('material-ui/lib/grid-list/grid-list');
+const GridList_1 = require('material-ui/GridList');
 const navtile_1 = require("./navtile");
 class NavTiles extends Component {
     render() {
@@ -9,7 +9,7 @@ class NavTiles extends Component {
         let tiles_ = tiles.map(function (data) {
             return (React.createElement(navtile_1.NavTile, {key: data.id, markup: data.content, help: data.help, tilecolors: tilecolors, system: system, route: data.route, transitionTo: transitionTo}));
         });
-        return (React.createElement(GridList, {style: style, children: tiles_, cols: tilecols, padding: padding, cellHeight: cellHeight}));
+        return (React.createElement(GridList_1.GridList, {style: style, children: tiles_, cols: tilecols, padding: padding, cellHeight: cellHeight}));
     }
 }
 exports.NavTiles = NavTiles;
