@@ -2,6 +2,8 @@
 const React = require('react');
 var { Component } = React;
 const injectTapEventPlugin = require('react-tap-event-plugin');
+const MuiThemeProvider_1 = require('material-ui/styles/MuiThemeProvider');
+const getMuiTheme_1 = require('material-ui/styles/getMuiTheme');
 injectTapEventPlugin();
 const redux_1 = require('redux');
 const react_redux_1 = require('react-redux');
@@ -27,7 +29,7 @@ if (!auth.isAuthenticated) {
 }
 class Main extends Component {
     render() {
-        return (React.createElement(react_redux_1.Provider, {store: store}, React.createElement("div", null, React.createElement(mainbar_1.MainBar, null), React.createElement("div", {style: { height: "64px" }}, " "), React.createElement("div", null, "THIS IS THE DEVELOPER'S VERSION OF THIS SITE (FOR PROTOTYPING)," + ' ' + "AND MAY CHANGE OR BREAK AT ANY TIME. ALSO, PLEASE NOTE THAT THE DATA" + ' ' + "HAS NOT BEEN VETTED."), routes_1.routes)));
+        return (React.createElement(MuiThemeProvider_1.default, {muiTheme: getMuiTheme_1.default()}, React.createElement(react_redux_1.Provider, {store: store}, React.createElement("div", null, React.createElement(mainbar_1.MainBar, null), React.createElement("div", {style: { height: "64px" }}, " "), React.createElement("div", null, "THIS IS THE DEVELOPER'S VERSION OF THIS SITE (FOR PROTOTYPING)," + ' ' + "AND MAY CHANGE OR BREAK AT ANY TIME. ALSO, PLEASE NOTE THAT THE DATA" + ' ' + "HAS NOT BEEN VETTED."), routes_1.routes))));
     }
 }
 exports.Main = Main;
