@@ -2,9 +2,7 @@
 const React = require('react');
 var { Component } = React;
 const react_redux_1 = require('react-redux');
-const Card = require('material-ui/lib/card/card');
-const CardText = require('material-ui/lib/card/card-text');
-const CardTitle = require('material-ui/lib/card/card-title');
+const Card_1 = require('material-ui/Card');
 class RegisterPendingClass extends Component {
     render() {
         let registerpendingpage = this;
@@ -18,7 +16,7 @@ class RegisterPendingClass extends Component {
                     React.createElement("div", null, React.createElement("p", null, "Thanks for registering, ", register.user.username, "!"), React.createElement("p", null, "An email has been sent to" + ' ' + "the email address you used to register. Please follow the instructions in this email" + ' ' + "to authenticate and complete your registration."))
                 :
                     React.createElement("div", null, React.createElement("p", null, "No registration data is available."));
-        return React.createElement(Card, {style: { margin: "5px" }}, React.createElement(CardTitle, {title: "Registration Pending", style: { paddingBottom: 0 }}), React.createElement(CardText, null, registerpending));
+        return React.createElement(Card_1.Card, {style: { margin: "5px" }}, React.createElement(Card_1.CardTitle, {title: "Registration Pending", style: { paddingBottom: 0 }}), React.createElement(Card_1.CardText, null, registerpending));
     }
 }
 function mapStateToProps(state) {
